ESP V0.2 software
ESP Timeclient changes

+++ Nixie Clock Software/ESP_Clock/ESP_Clock.ino
@@ -161,7 +161,7 @@ void loop() {
       }
       else if(command == "GetEpoch")
       {
-        long int currentEpoch = timeClient.getCurrentEpoch();
+        uint32_t currentEpoch = timeClient.getCurrentEpoch();
         Serial.print(String(currentEpoch) + ";");
       }
       else if(command == "GetSignalStrength")
